requirements.txt
requests>=2.31.0
pytz>=2023.3
PyYAML>=6.0.1